+++ src/main/java/com/github/danielflower/mavenplugins/release/Reactor.java
@@ -1,5 +1,6 @@
 package com.github.danielflower.mavenplugins.release;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.model.Dependency;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
@@ -28,7 +29,7 @@ public class Reactor {
         return modulesInBuildOrder;
     }
 
-    public static Reactor fromProjects(Log log, LocalGitRepo gitRepo, MavenProject rootProject, List<MavenProject> projects, Long buildNumber, List<String> modulesToForceRelease, NoChangesAction actionWhenNoChangesDetected, ResolverWrapper resolverWrapper, VersionNamer versionNamer) throws ValidationException, GitAPIException, MojoExecutionException {
+    public static Reactor fromProjects(Log log, LocalGitRepo gitRepo, MavenProject rootProject, List<MavenProject> projects, String buildNumber, List<String> modulesToForceRelease, NoChangesAction actionWhenNoChangesDetected, ResolverWrapper resolverWrapper, VersionNamer versionNamer) throws ValidationException, GitAPIException, MojoExecutionException {
         DiffDetector detector = new TreeWalkingDiffDetector(gitRepo.git.getRepository());
         List<ReleasableModule> modules = new ArrayList<ReleasableModule>();
 
@@ -42,14 +43,14 @@ public class Reactor {
             List<AnnotatedTag> previousTagsForThisModule = annotatedTagFinder.tagsForVersion(gitRepo.git, artifactId, versionWithoutBuildNumber);
 
 
-            Collection<Long> previousBuildNumbers = new ArrayList<Long>();
+            Collection<String> previousBuildNumbers = new ArrayList<>();
             if (previousTagsForThisModule != null) {
                 for (AnnotatedTag previousTag : previousTagsForThisModule) {
                     previousBuildNumbers.add(previousTag.buildNumber());
                 }
             }
 
-            Collection<Long> remoteBuildNumbers = getRemoteBuildNumbers(gitRepo, artifactId, versionWithoutBuildNumber, versionNamer);
+            Collection<String> remoteBuildNumbers = getRemoteBuildNumbers(gitRepo, artifactId, versionWithoutBuildNumber, versionNamer);
             previousBuildNumbers.addAll(remoteBuildNumbers);
 
             VersionName newVersion = versionNamer.name(project.getVersion(), buildNumber, previousBuildNumbers);
@@ -145,14 +146,14 @@ public class Reactor {
         return modelId[0].equals(module.getGroupId()) && modelId[1].equals(module.getArtifactId());
     }
 
-    private static Collection<Long> getRemoteBuildNumbers(LocalGitRepo gitRepo, String artifactId, String versionWithoutBuildNumber, VersionNamer versionNamer) throws GitAPIException {
+    private static Collection<String> getRemoteBuildNumbers(LocalGitRepo gitRepo, String artifactId, String versionWithoutBuildNumber, VersionNamer versionNamer) throws GitAPIException {
         Collection<Ref> remoteTagRefs = gitRepo.allTags();
-        Collection<Long> remoteBuildNumbers = new ArrayList<Long>();
+        Collection<String> remoteBuildNumbers = new ArrayList<>();
         String tagWithoutBuildNumber = artifactId + "-" + versionWithoutBuildNumber;
         AnnotatedTagFinder annotatedTagFinder = new AnnotatedTagFinder(versionNamer);
         for (Ref remoteTagRef : remoteTagRefs) {
             String remoteTagName = remoteTagRef.getName();
-            Long buildNumber = annotatedTagFinder.buildNumberOf(tagWithoutBuildNumber, remoteTagName);
+            String buildNumber = annotatedTagFinder.buildNumberOf(tagWithoutBuildNumber, remoteTagName);
             if (buildNumber != null) {
                 remoteBuildNumbers.add(buildNumber);
             }
@@ -190,20 +191,25 @@ public class Reactor {
         try {
             if (previousTagsForThisModule.size() == 0) return null;
             boolean hasChanged = detector.hasChangedSince(relativePathToModule, project.getModel().getModules(), previousTagsForThisModule);
-            return hasChanged ? null : tagWithHighestBuildNumber(previousTagsForThisModule);
+            return hasChanged ? null : findTagWithHighestBuildNumber(previousTagsForThisModule);
         } catch (Exception e) {
             throw new MojoExecutionException("Error while detecting whether or not " + project.getArtifactId() + " has changed since the last release", e);
         }
     }
 
-    private static AnnotatedTag tagWithHighestBuildNumber(List<AnnotatedTag> tags) {
-        AnnotatedTag cur = null;
+    // Ignores any tags with non-numeric build numbers and finds the one with the largest build number among the others.
+    // The tags are guaranteed to be for the same /business version/ i.e. it's likely a small list.
+    private static AnnotatedTag findTagWithHighestBuildNumber(List<AnnotatedTag> tags) {
+        AnnotatedTag tagWithHighestBuildNumber = null;
         for (AnnotatedTag tag : tags) {
-            if (cur == null || tag.buildNumber() > cur.buildNumber()) {
-                cur = tag;
+            if(StringUtils.isNumeric(tag.buildNumber())) {
+                if (tagWithHighestBuildNumber == null ||
+                    Long.parseLong(tag.buildNumber()) > Long.parseLong(tagWithHighestBuildNumber.buildNumber())) {
+                    tagWithHighestBuildNumber = tag;
+                }
             }
         }
-        return cur;
+        return tagWithHighestBuildNumber;
     }
 
     public ReleasableModule findByLabel(String label) {