+++ src/test/java/com/github/danielflower/mavenplugins/release/AnnotatedTagFinderTest.java
@@ -22,22 +22,22 @@ public class AnnotatedTagFinderTest {
     public void findsTheLatestCommitWhereThereHaveBeenNoBranches() throws Exception {
         TestProject project = TestProject.independentVersionsProject();
 
-        AnnotatedTag tag1 = saveFileInModule(project, "console-app", "1.2", 3);
-        AnnotatedTag tag2 = saveFileInModule(project, "core-utils", "2", 0);
-        AnnotatedTag tag3 = saveFileInModule(project, "console-app", "1.2", 4);
+        AnnotatedTag tag1 = saveFileInModule(project, "console-app", "1.2", "3");
+        AnnotatedTag tag2 = saveFileInModule(project, "core-utils", "2", "0");
+        AnnotatedTag tag3 = saveFileInModule(project, "console-app", "1.2", "4");
 
         assertThat(annotatedTagFinder.tagsForVersion(project.local, "console-app", "1.3"), hasSize(0));
         assertThat(annotatedTagFinder.tagsForVersion(project.local, "console-app", "1.2"), containsInAnyOrder(tag1, tag3));
         assertThat(annotatedTagFinder.tagsForVersion(project.local, "core-utils", "2"), contains(tag2));
     }
 
-    static AnnotatedTag saveFileInModule(TestProject project, String moduleName, String version, long buildNumber) throws IOException, GitAPIException {
+    static AnnotatedTag saveFileInModule(TestProject project, String moduleName, String version, String buildNumber) throws IOException, GitAPIException {
         project.commitRandomFile(moduleName);
         String nameForTag = moduleName.equals(".") ? "root" : moduleName;
         return tagLocalRepo(project, nameForTag + "-" + version + versionNamer.getDelimiter() + buildNumber, version, buildNumber);
     }
 
-    private static AnnotatedTag tagLocalRepo(TestProject project, String tagName, String version, long buildNumber) throws GitAPIException {
+    private static AnnotatedTag tagLocalRepo(TestProject project, String tagName, String version, String buildNumber) throws GitAPIException {
         AnnotatedTag tag = AnnotatedTag.create(tagName, version, buildNumber);
         tag.saveAtHEAD(project.local);
         return tag;
@@ -54,10 +54,10 @@ public class AnnotatedTagFinderTest {
     @Test
     public void returnsMultipleTagsOnASingleCommit() throws IOException, GitAPIException, MojoExecutionException {
         TestProject project = TestProject.independentVersionsProject();
-        saveFileInModule(project, "console-app", "1.2", 1);
-        AnnotatedTag tag1 = tagLocalRepo(project, "console-app-1.1.1.1", "1.1.1", 1);
-        AnnotatedTag tag3 = tagLocalRepo(project, "console-app-1.1.1.3", "1.1.1", 3);
-        AnnotatedTag tag2 = tagLocalRepo(project, "console-app-1.1.1.2", "1.1.1", 2);
+        saveFileInModule(project, "console-app", "1.2", "1");
+        AnnotatedTag tag1 = tagLocalRepo(project, "console-app-1.1.1.1", "1.1.1", "1");
+        AnnotatedTag tag3 = tagLocalRepo(project, "console-app-1.1.1.3", "1.1.1", "3");
+        AnnotatedTag tag2 = tagLocalRepo(project, "console-app-1.1.1.2", "1.1.1", "2");
         List<AnnotatedTag> annotatedTags = annotatedTagFinder.tagsForVersion(project.local, "console-app", "1.1.1");
         assertThat(annotatedTags, containsInAnyOrder(tag1, tag2, tag3));
     }
@@ -65,7 +65,7 @@ public class AnnotatedTagFinderTest {
     @Test
     public void versionNamerCaresNotForOrderOfTags() throws ValidationException {
         VersionNamer versionNamer = new VersionNamer();
-        VersionName name = versionNamer.name("1.1.1", null, asList(1L, 3L, 2L));
+        VersionName name = versionNamer.name("1.1.1", null, asList("1", "3", "2"));
         assertThat(name.releaseVersion(), equalTo("1.1.1.4"));
     }
 
@@ -76,15 +76,15 @@ public class AnnotatedTagFinderTest {
         // This should actually be implemented with a parameterized JUnit test but since this class isn't set up for
         // that yet we'll use our own data-driven harness.
         Arrays.asList(new Object[][]{
+            new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0-1", "1"},    // proper match with expected delimiter
+            new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0-A", "A"},    // non-numeric build number
             new Object[]{"libs-1.0.0", refPrefix + "foo-1.0.0-1", null},    // artifact ID does not match -> expect null
-            new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0-1", 1L},     // proper match with expected delimiter
             new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0.1", null},   // no delimiter match
             new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0@1", null},   // no delimiter match
-            new Object[]{"libs-1.0.0", refPrefix + "libs-1.0.0-A", null}    // format exception on build number
         })
         .forEach(dataSet -> {
             final Object expectedBuildNumber = dataSet[2];
-            final Long buildNumber = tagFinder.buildNumberOf(dataSet[0].toString(), dataSet[1].toString());
+            final String buildNumber = tagFinder.buildNumberOf(dataSet[0].toString(), dataSet[1].toString());
             assertThat(buildNumber, is(expectedBuildNumber));
         });
     }