+++ src/test/java/scaffolding/MvnRunner.java
@@ -1,12 +1,14 @@
 package scaffolding;
 
 import org.apache.commons.io.filefilter.DirectoryFileFilter;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.shared.invoker.*;
 
 import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
 import java.io.PrintStream;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
@@ -82,6 +84,8 @@ public class MvnRunner {
 
         int exitCode;
         try {
+            System.out.println(String.format("Running mvn in '%s' with profile(s) '%s' and arguments %s.",
+                workingDir, StringUtils.defaultString(profile), Arrays.toString(arguments)));
             InvocationResult result = invoker.execute(request);
             exitCode = result.getExitCode();
         } catch (Exception e) {
@@ -93,6 +97,8 @@ public class MvnRunner {
             throw new MavenExecutionException(exitCode, output);
         }
 
+        System.out.println("That produced the following output.\n");
+        output.forEach(System.out::println);
         return output;
     }